+++ lustre/utils/lhsmtool_posix.c
@@ -1675,7 +1675,7 @@ static int ct_max_sequence(void)
 		sprintf(path + strlen(path), "/%04x", subseq);
 	}
 
-	printf("max_sequence: %016Lx\n", seq);
+	printf("max_sequence: %016Lx\n", (long long unsigned int)seq);
 
 	return 0;
 }
